Improve command line interface and database in offline mode
The command line is now in interactive mode to avoid reconnecting for
every command. The parser could be simplified while doing this.
The discover module in gossip is catching network errors for cases where
the network is down and sending a PING is useless.
Furthermore some modifications in the database is done to provide
transitions in the protocol.