Add store action to cli and improve performance
The command line interface has a new query named `store` which can be
used to store an arbitrary music file in the database. It will read the
id3tag convert it to Opus and create a new entry in the database. At the
same time the music-container save routine was improved to support
large audio file (longer than 1h).