Split database view into (Reader, Writer, Files)
This commit simplifies the internal structure of the database module.
It does so by splitting the old `View` struct into `Reader`, `Writer`
and `Files` structs. Furthermore the `Stream` trait is now implemented
for the database, instead of a internal thread which handles the
eventloop. This means that an `Instance` has now to be executed on the event
loop.